Refactoring Ubuntu-20.04-BasicSetup.sh

+++ Ubuntu-20.04-BasicSetup.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+CONDA_BIN=~/miniconda3/condabin/conda
+INSTALLATION_LOG_DIR=logs
+
 sudo apt update && sudo apt upgrade -y
 
 # Install basic packages
@@ -32,21 +35,18 @@ install_python3_pip() {
 }
 
 install_miniconda() {
-  mkdir -p ~/miniconda3
-  wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O ~/miniconda3/miniconda.sh
-  bash ~/miniconda3/miniconda.sh -b -u -p ~/miniconda3
-  rm -rf ~/miniconda3/miniconda.sh
-  ~/miniconda3/bin/conda init bash
-  ~/miniconda3/bin/conda config --set auto_activate_base false
+  if [ ! -f "$CONDA_BIN" ]; then
+    mkdir -p ~/miniconda3
+    wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O ~/miniconda3/miniconda.sh
+    bash ~/miniconda3/miniconda.sh -b -u -p ~/miniconda3
+    rm -rf ~/miniconda3/miniconda.sh
+    ~/miniconda3/bin/conda init bash
+    ~/miniconda3/bin/conda config --set auto_activate_base false
+  fi
 }
 
 install_python_conda_packages() {
-  ~/miniconda3/bin/conda install -y -c conda-forge \
-    jupyterlab \
-    notebook \
-    voila \
-    pandas \
-    ipython
+  eval "${CONDA_BIN} install -y -c conda-forge jupyterlab notebook voila pandas ipython"
 }
 
 install_php7() {
@@ -58,13 +58,15 @@ install_php7() {
 }
 
 install_composer() {
-  curl -sS https://getcomposer.org/installer -o composer-setup.php
-  HASH=`curl -sS https://composer.github.io/installer.sig`
-  php -r "if (hash_file('SHA384', 'composer-setup.php') === '$HASH') { echo 'Installer verified'; } else { echo 'Installer corrupt'; unlink('composer-setup.php'); } echo PHP_EOL;"
-  sudo php composer-setup.php --install-dir=/usr/local/bin --filename=composer
-  
-  rm -f composer-setup.php
-  export_composer_binpaths
+  if [ ! -f /usr/local/bin/composer ]; then
+    curl -sS https://getcomposer.org/installer -o composer-setup.php
+    HASH=`curl -sS https://composer.github.io/installer.sig`
+    php -r "if (hash_file('SHA384', 'composer-setup.php') === '$HASH') { echo 'Installer verified'; } else { echo 'Installer corrupt'; unlink('composer-setup.php'); } echo PHP_EOL;"
+    sudo php composer-setup.php --install-dir=/usr/local/bin --filename=composer
+    
+    rm -f composer-setup.php
+    export_composer_binpaths
+  fi
 }
 
 install_nvm() {
@@ -123,19 +125,40 @@ install_gradle() {
 }
 
 install_dotnet_sdk() {
-  wget https://packages.microsoft.com/config/ubuntu/20.04/packages-microsoft-prod.deb -O packages-microsoft-prod.deb
-  sudo dpkg -i packages-microsoft-prod.deb
+  DOTNET_BIN=/usr/bin/dotnet
 
-  sudo apt-get update; \
-    sudo apt-get install -y apt-transport-https && \
-    sudo apt-get update && \
-    sudo apt-get install -y dotnet-sdk-3.1
+  if [ ! -f "$DOTNET_BIN" ]; then
+    wget https://packages.microsoft.com/config/ubuntu/20.04/packages-microsoft-prod.deb -O packages-microsoft-prod.deb
+    sudo dpkg -i packages-microsoft-prod.deb
 
-  rm -f packages-microsoft-prod.deb
+    sudo apt-get update; \
+      sudo apt-get install -y apt-transport-https && \
+      sudo apt-get update && \
+      sudo apt-get install -y dotnet-sdk-3.1
+
+    rm -f packages-microsoft-prod.deb
+  fi
 }
 
 install_dotnet_runtime() {
   sudo apt-get install -y dotnet-runtime-3.1
+}
+
+install_geckodriver() {
+  GECKODRIVER_VERSION=0.27.0
+  GECKODRIVER_OS_ARCH=linux64
+  GECKODRIVER_TAR_FILENAME="geckodriver-v${GECKODRIVER_VERSION}-${GECKODRIVER_OS_ARCH}.tar.gz"
+  GECKODRIVER_BIN_FILENAME=geckodriver
+  GH_URL="https://github.com/mozilla/geckodriver/releases/download/v${GECKODRIVER_VERSION}/$GECKODRIVER_TAR_FILENAME"
+
+  if [[ -z `which $GECKODRIVER_BIN_FILENAME` ]]; then
+    wget "$GH_URL"
+    tar -xvf "$GECKODRIVER_TAR_FILENAME"
+    sudo mv "$GECKODRIVER_BIN_FILENAME" /usr/local/bin/
+    rm -f "$GECKODRIVER_TAR_FILENAME"
+  else
+    echo "Mozilla Geckodriver already installed."
+  fi
 }
 
 setup_direnv_bash() {
@@ -153,49 +176,52 @@ export LIBGL_ALWAYS_INDIRECT=1
 EOF
 }
 
-mkdir -p Ubuntu-20.04-BasicSetup
+mkdir -p "$INSTALLATION_LOG_DIR"
 
 echo "Installing Miniconda3 Package Tool..."
-install_miniconda > Ubuntu-20.04-BasicSetup/miniconda_install.log
+install_miniconda > "$INSTALLATION_LOG_DIR/miniconda_install.log"
 
 echo "Installing Python Common Packages..."
-install_python_conda_packages > Ubuntu-20.04-BasicSetup/python_conda_packages_install.log
+install_python_conda_packages > "$INSTALLATION_LOG_DIR/python_conda_packages_install.log"
 
 echo "Installing PHP 7..."
-install_php7 > Ubuntu-20.04-BasicSetup/php7_install.log
+install_php7 > "$INSTALLATION_LOG_DIR/php7_install.log"
 
 echo "Installing Composer Package Tool..."
-install_composer > Ubuntu-20.04-BasicSetup/composer_install.log
+install_composer > "$INSTALLATION_LOG_DIR/composer_install.log"
 
 echo "Installing NVM (NodeJS Version Manager)..."
-install_nvm > Ubuntu-20.04-BasicSetup/nvm_install.log
+install_nvm > "$INSTALLATION_LOG_DIR/nvm_install.log"
 
 echo "Installing NodeJS v12 LTS..."
-install_nodejs_12_lts > Ubuntu-20.04-BasicSetup/nodejs12_lts_install.log
+install_nodejs_12_lts > "$INSTALLATION_LOG_DIR/nodejs12_lts_install.log"
 
 echo "Installing Yarn..."
-install_yarn > Ubuntu-20.04-BasicSetup/yarn_install.log
+install_yarn > "$INSTALLATION_LOG_DIR/yarn_install.log"
 
 echo "Installing Java JRE..."
-install_java_jre > Ubuntu-20.04-BasicSetup/java_jre_install.log
+install_java_jre > "$INSTALLATION_LOG_DIR/java_jre_install.log"
 
 echo "Installing Java JDK..."
-install_java_jdk > Ubuntu-20.04-BasicSetup/java_jdk_install.log
+install_java_jdk > "$INSTALLATION_LOG_DIR/java_jdk_install.log"
 
 echo "Installing SDKMAN..."
-install_sdkman > Ubuntu-20.04-BasicSetup/sdkman_install.log
+install_sdkman > "$INSTALLATION_LOG_DIR/sdkman_install.log"
 
 echo "Installing Gradle..."
-install_gradle > Ubuntu-20.04-BasicSetup/gradle_install.log
+install_gradle > "$INSTALLATION_LOG_DIR/gradle_install.log"
 
 echo "Installing .NET Core 3.1 SDK..."
-install_dotnet_sdk > Ubuntu-20.04-BasicSetup/dotnet_sdk_install.log
+install_dotnet_sdk > "$INSTALLATION_LOG_DIR/dotnet_sdk_install.log"
 
 echo "Installing .NET Core 3.1 Runtime..."
-install_dotnet_runtime > Ubuntu-20.04-BasicSetup/dotnet_runtime_install.log
+install_dotnet_runtime > "$INSTALLATION_LOG_DIR/dotnet_runtime_install.log"
+
+echo "Installing Mozilla Geckodriver..."
+install_geckodriver > "$INSTALLATION_LOG_DIR/geckodriver_install.log"
 
 echo "APT Autoremove..."
-apt_autoremove > Ubuntu-20.04-BasicSetup/apt_autoremove.log
+apt_autoremove > "$INSTALLATION_LOG_DIR/apt_autoremove.log"
 
 echo "Setup direnv..."
 setup_direnv_bash
